Refactor JNI for speaker identification
