+++ android/SherpaOnnxAudioTagging/app/src/main/java/com/k2fsa/sherpa/onnx/audio/tagging/AudioTagging.kt
@@ -2,8 +2,6 @@ package com.k2fsa.sherpa.onnx
 
 import android.content.res.AssetManager
 
-const val TAG = "sherpa-onnx"
-
 data class OfflineZipformerAudioTaggingModelConfig(
     var model: String = "",
 )
